requirements.txt
attrs==21.2.0
greenlet==1.1.1
iniconfig==1.1.1
packaging==21.0
playwright==1.14.1
pluggy==0.13.1
py==1.10.0
pyee==8.2.2
pyparsing==2.4.7
pytest==6.2.4
toml==0.10.2
websockets==9.1
colorama~=0.4.4
pip~=21.1.2
wheel~=0.36.2
pytz~=2021.1
Django~=3.1.4
sqlparse~=0.4.1
asgiref~=3.4.1
atomicwrites~=1.4.0
setuptools~=57.4.0
allure-pytest==2.9.43
